Add tab searching
All the open tabs can now be searched. The searching is done in
background script and result is sent to the iframe. The search bar
shows the matched tab titles.

Fuzzy search is used for search logic using fuse.js. This makes it
easy to search tabs by just typing a few letters from the title.