Add Themes 🎨
Themes plugin is added which allows the user to search and select
a theme for UnTab.
Implementation of the themes are done via CSS variables which are
defined in `src/themes/themes.css` file. New themes can be easily
added by defining new colors in the `themes.css` file and adding
an item in the `themes` plugin.

Themes that are currently present:
- Dark
- Light
- Coffee

Fixes #9