Refactor Tab switching to a plugin
UnTab started out as a tool to search and switch between tabs.
With its plugin system expanded over time, it is now possible to
migrate the tab switching to a plugin on its own without much
changes!

This makes UnTab a shell for running a bunch of plugins and
keeping users productive in their browser usage.